Fix gradlew bouncer output
The escaped $ wound up in the source, and broke running it:

```
Starting a Gradle Daemon, 1 incompatible Daemon could not be reused, use --status for details

> Configure project :SharedCode
Kotlin Multiplatform Projects are an experimental feature.

FAILURE: Build failed with an exception.

* What went wrong:
Task '$@' not found in root project 'HandsOnKMP'.

* Try:
Run gradlew tasks to get a list of available tasks. Run with --stacktrace option to get the stack trace. Run with --info or --debug option to get more log output. Run with --scan to get full insights.

* Get more help at https://help.gradle.org

BUILD FAILED in 4s
Command PhaseScriptExecution failed with a nonzero exit code
```